requirements.txt
# Core dependencies
celery==5.3.6
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Audio processing
torch==2.2.0
torchaudio==2.2.0
SpeechRecognition==3.10.1
numpy==1.26.3

# API and HTTP
requests==2.31.0
python-dotenv==1.0.1

# Database
pymongo==4.6.1

# AI/ML
google-generativeai==0.3.2

# Testing
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0